requirements.txt
flask
flask_sqlalchemy
flask_marshmallow
marshmallow-sqlalchemy
flask_cors
flask_mail
apscheduler
python-dotenv
